pyproject.toml
[project]
name = "pySMC"
version = "0.1.0"
description = "Probabilistic sampling, cases, and surrogate modeling utilities"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "yuma" }]

dependencies = [
  # Core scientific computing
  "numpy>=1.26",
  "pandas>=2.1",
  "scipy>=1.11",
  
  # Machine learning and optimization
  "scikit-learn>=1.3",
  "jax>=0.6.2",
  "optax>=0.2.6",
  
  # Visualization
  "matplotlib>=3.8",  # Relaxed from 3.10.7 for better compatibility
  
  # MCP server framework
  "mcp[cli]>=1.20.0",
  "fastmcp>=0.1.0",
  
  # Web interface
  "streamlit>=1.28.0",
  
  # LLM integration
  "openai>=1.0.0",
  
  # Security and authentication
  "python-dotenv>=1.0.0",
  "cryptography>=41.0.0",
  "pyotp>=2.9.0",
  "qrcode>=7.4.2",
  "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",  # Code coverage
  "ruff>=0.6",
  "black>=23.0",      # Code formatter
  "mypy>=1.0",        # Type checking
]

jupyter = [
  "jupyter>=1.0",
  "ipykernel>=6.0",
  "ipywidgets>=8.0",
]

viz = [
  "seaborn>=0.12",    # Enhanced plotting
  "plotly>=5.0",      # Interactive plots
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


